Fixing references used by the performance projects
This fixes issue #991

The TypeLoadException was incorrectly reporting that a method had no body when in reality it was unable to locate one of the dependencies. It's not clear why this exception was reported, but the error stops showing after fixing the project references (both in the csproj and in project.json), followed by a .\build initialize and a clean build of the EntityFramework solution.

The performance projects are now running properly under net45.