Remove incomplete UseXxx service methods
These were intended as a form of sugar for replacing services. However, they have not been updated with new services, are used inconsistently, and don't seem to provide real value over just using the normal DI methods. We already know that we need to consider the common extensibility story with some form of sugar, but I don't think this is it, so cleaning this out for now.