Create abstraction for state entry metadata services
This is part of Issue #641 which is about cleaning up the use of DbContextConfiguration. DbContextConfiguration is intended to help resolve dynamic services--that is, services for which the actual type/instance of service to use depends on the current context configuration. However, it has gradually spread throughout the code as a general purpose service locator. This causes dependencies to be hidden and creates a lot of coupling to DbContextConfiguration throughout the code, so I am making a set of changes to help fix this.

A lot of the services used by a StateEntry are about accessing and using metadata related to the entity type. These have been abstracted out so that the StateEntry does not need to use the DbContextConfiguration service locator for these things.