Making Ignore behave the same independently of the order relative to other configuration.
If a property or entity was added and ignored explicitly an exception will be thrown.
If a property or entity was added and ignored through conventions or data annotations the property will be present on the entity.
If a property is ignored and removed after conventions or data annotations added keys, foreign keys or indexes referencing it they will be removed as well.
The same applies to entity types and foreign keys referencing them.