Change casing of SQLite classes (Issue #1047)
Current casing matches the external SQLite ADO.NET provider that we don't use at all, but does not conform to any naming standards. A few of us talked and came to the conclusion that Sqlite would be better. Explaining: SQLite is not two words--i.e. it's not SQ Lite or SQL Ite, which means that SqLite is wrong and SQLite is wrong. So it must be one (weird) word and hence casing should be Sqlite.