Update handling of snapshots and ensure DetectChanges is only used when required (Issue #731)
The logic around when to snapshot original values was written before a lot of the relationship fixup and DetectChanges code. These changes revisit that code to improve and test the logic for determined if and when snapshots are required.

We make eager snapshots (as soon as the entity is tracked) for original values and relationship properties and navigations when the entity does not implement INotifyPropertyChanging. In addition, we take a collection snapshot for any collection that does not implement INotifyCollectionChanged. However, since INotifyCollectionChanged is not actually supported yet we currently treat all collections as if they do not implement this interface. See issue #445.

DetectChanges is needed for any entity that does not implement INotifyPropertyChanged, and also if an entity has a collection that does not implement INotifyCollectionChanged. As above, it is currently assumed that no collection does because of issue #445.