Fix to #1528 - Exception thrown for multi-level include dotting over one-one navigation into one to many deeper inside the include chain
Problem was that when we tried to extract list of PropertyInfos that include chain consists of, we were doing it incorrectly - always assuming it would be a simple property (or list of those inside anonymous type).
Fix is to use a proper method that expects complex properties instead. CR: Andrew