Fix to multiple include bugs - #1433, #1478, #1487 #1488
Problem was that if a query was executed multiple times, or after another query, some of the entities were stored in state manager and not in the reader buffer. However when accessing PK/FK we were always using the buffered readers. This could cause multitude of errors, e.g buffer being empty, trying to access element of the reader that doesn't exist, is of a wrong type, or worst case - cause data corruption.

Fix is to mark first element of the buffer for a given entity with a null value, if that entity is already in state manager - later, when we try to retrieve the keys we check for that null, and if we find it we use state manager instead.