Fix to #1251 - join + orderby + count produces invalid sql
Problem was that we would procude queries like:
SELECT COUNT(*)
FROM [Customer] AS [c]
ORDER BY [c].[Id]

which were invalid, because column on which you order should be projected, and we project COUNT(*) instead. However, ordering is redundant when querying for count and could be removed.

Fix is to remove any orderings when querying for count, just like we already do for some other result operators (e.g. Any)