Make value generators share fewer features and allow scoped value generators
Conceptually a value generator is a singleton service that should persist across scopes. This is the way they were previously implemented. However, this makes it impossible for a value generators to use scoped services, such as the connection, unless those services are passed to the generator instance. This change allows scoped value generators to be created such that they can use scoped services. This allows a generator to be created that can use, for example, the connection. However, care must then be taken to ensure that the state of the generator persists across scopes (for which a cache is provided.)

In addition, two of the reusable features of value generators--the ability to cache by something other than a property, and the ability to use a pool of generators--have been removed. These are now built into the HiLo/Sequence generator, which was the only generator currently using these features. Other generators cannot be pooled and are always cached by property unless they also implement these services inside their own code.