fix to #1642 -Query :: Bad sql generated for query with bool parameter in filter
Error is for queries that are using bool parameter (closure) in a filter query. We would produce query like: SELECT c.Id FROM Customers as c WHERE @prm, which is invalid.

Fix is to apply similar compensation as we do for constant bools in filter, to produce query like: SELECT c.Id FROM Customers as c WHERE @prm = 1