Cleanup DbContextOptions and factor our Builder
This change cleans up the DbContextOptions classes, the way they are built and handled, and the way extensions interact. There is now a DbContextOptionsBuilder class that is used in AddEntityFramework and OnConfiguring. This is used to create DbContextOptions objects which can be passed to the DbContext constructor and/or resolved from DI. From the app perspective everything works pretty much the same except that the DbContextOptions resolved from DI is no longer cloned before being passed to OnConfiguring. However, if OnConfiguring then does anything to change the options it will effectively result in the options being cloned at this time. The DbContextOptions themselves are effectively immutable.

All additional options are now added through extensions, and the way extensions add added/updated has been simplified a bit to have easier APIs and to decouple more cleanly from the raw options that are handled by DbContextOptions itself.

Auto configuring of a single data store now happens right after OnConfiguring is called, rather than later when a data store is selected. This is cleaner and means data store selection doesn't need to mutate options.

It is now possible to depend on non-generic DbContextOptions in your DbContext constructor, but a generic DbContextOptions for your context type must be used. Non-generic DbContextOptions is now abstract.

DbContextOptions parsing is now performed by a DI service.