Fix to #1747 - Multi-select with Contains() does not generate SQL
Fix is to translate Where(c => [list_of_constants].Constains(c.SomeProperty)) to
c.SomeProperty IN ( [list_of_constants)

We also translate !Contains... into NOT IN. Relinq converts Contains call into subquery, so we are detecting this pattern in the FilteringExpressionTreeVisitor and translate it accordingly. Argument to contains in this case will always be in the form of parameter. Parameter values are wired in during SQL generation.
Refactored EqualityPredicateOptimizer to factor in In/NotIn nodes created via Contains in the optimizations. Removed

Additional refactoring: NotInExpression and InExpressionBase, recognizing pattern NotExpression(InExpression(..)) instead.

CR: Andrew