Use generic ReadValue in update pipeline and key factories
This change allows cached delegates to be created that will call the generic IValueReader.ReadValue method using the type of the property. This means that when a RelationalTypedValueReader is used the generic GetFieldValue method will be called on the underlying data reader, thereby giving the provider the contextual information needed to create objects of the correct type in cases where this cannot be determined by the database metadata alone.

This mechanism is being used in the update pipeline and for key generation. It is not yet being used in query or for reading shadow state properties.