Additional fix to Issue 1747 - Multi-select with Contains() does not generate SQL
Problem is we were not handling the case where user would pass empty collection as argument to contains - we would produce invalid sql.
Fix is to check for number of arguments to InExpression before producing a relevant sql.
This has to happen during sql generation, since we need to know the values of the parameters.

CR: Andrew