:apple: Workarounds to support Mono
Two workarounds needed...

**ThenInclude method location**
IsConstructedGenericType is returning different results on Mono so adjusting logic to something that works.
New logic for the collection overload just uses a "not" of the logic to find the reference overload.

**Remove use of DbDataReader.GetFieldValue<T>**
There is a bug in Mono where `GetFieldValue<T>` is overridden in `SqlDataReader` to throw even though there is now a base implementation that would work.
I've opened an issue against Mono to fix this - https://bugzilla.xamarin.com/show_bug.cgi?id=28662
This is just a temporary workaround that will need to be done in a better way since we want the generic to help with type detection on SQLite and possibly other providers.