Extend read-only semantics per store generation discussion
Note: the naming of the concepts and whether or not they should be in an enum can be discussed in API review.

This change adds the concept of being read-only before saving a new entity as well as the concept of being read-only when updating a new entity. It also moves the check to happen as part of SaveChanges so that properties can go through a temporary state of being modified. Key values cannot go through this temporary state since the state manager relies on their values not changing.

Also added functional tests for all combinations of read-only flags and store generated patterns.