Fix to #1927 - Query :: data corruption for queries with filter and && operator where one side is client eval and another is server eval
Problem is that for predicates with && operator, for which one side was client eval and the other was server eval we would only process the server eval part. This could lead to data corruption (i.e. query returning more rows than intended)

Fix is to handle top level && operation differently - splitting predicate into two Where operations, one on client and one on server. This is done only for && operation and only for top level. If client eval is needed at any point and the top level expression is not AndAlso, the entire predicate will be evaluated on the client.

CR: Andrew