Allow generic GetFieldValue to be used in query
This change allows a value reader to be code genned that eagerly reads all values passing required type information to the ADO.NET provider. This requires that the types needed be captured/calculated in query. The code gennned delegates are cached and reused across the stack, not just for query.

Some additional cleanup/optimization and testing still to come. Primarily tested by running all the tests using this value reader on SQL Server, even though it is not currently used by SQL Server by default.

One test (Take_with_single_select_many) is failing with the new reader, will investigate more and/or file a bug.