Fix to #1838 - Query :: implement CLR null semantics for query predicates
Problem is that database semantics for null equality are different from c#, resulting in unexpected results for some queries (involving null value comparisons)

Fix is to compensate this by introducing additional terms in the predicate, effectively converting 3-valued logic into 2-valued logic.

a == b -> [(a == b) && (a != null && b != null)] || (a == null && b == null))
a != b -> [(a != b) || (a == null || b == null)] && (a != null || b != null)]

also added transformations for cases when one of the operands is negated

!a == b -> [(a != b) && (a != null && b != null)] || (a == null && b == null)
!a != b -> [(a == b) || (a == null || b == null)] && (a != null || b != null)

Some terms can be removed if one of the terms is not nullable, e.g:

nullableA == nonNullableB -> (nullableA == nonNullableB) && (nullableA != null)

We do additional optimization for simple predicate of a == b where a and b are nullable:
a == b -> (a == b) || (a == null && b == null)

This is still 3-valued result, true being returned when 'a' and 'b' are in c# terms, and either false or NULL when they are not equal.
This optimization can only be applied if the predicate does not contain any NotEqual or Not operations.

CR: Andrew