Fix to #2026 - consider different expansion for complex equality predicates (e.g. (a == b) == c) for some cases.
This is a performance/sqlgen improvement.
Currently expressions like (a == b) == (c) are rewritten into case statements:
(CASE WHEN a == b then 1 ELSE 0) == (CASE WHEN c == 1 then 1 ELSE 0)

This is not necessary as we only need to expand the part that is complex (a==b) and we can leave the simple part (c) intact.