Remove ForRelational
This change promotes the general-purpose relational extensions to be first class on the fluent API meaning that there is no longer a ForRelational call. This reflects the idea that relational is special and that the experience should be optimized as such. All other provider extensions, including extensions specific to a given relational provider, should continue use the existing pattern--ForSqlServer, etc.