Fix to #1803 - Remove CaseExpression
This change removes CaseExpression that has been introduced in EF, and used ConditionalExpression instead. Using this instead of Case/Switch doesn't allow for the full expressiveness of T-SQL CASE WHEN THEN, but it allows creating c# conditional (?:), as well as simple CASE-WHEN-THEN-ELSE that we use currently for our queries.

CR: Andrew