Update relational type mapping
This change updates the relational type mapping code such that:
- Specifying store types (e.g. nvarchar(128)) are recognized and parameters are created appropriately
- Specifying MaxLength is recognized and used in Migrations and in parameters
- DbType is only specified where needed--for example for DateTime2 on SQL Server

The update pipeline and Migrations use the type mapping. Query does not use it yet--it just creates ad-hoc parameters)

SQL Server has been fully tested. SQLite is not yet tested.

Using provider-specific type enums is not supported--we'll do that if we find a case that needs it.

Type conversions are not yet supported--the ADO.NET provider has to be able to work with your property type and your store type without going through an intermediate CLR type.