Add functional tests for SQLite type mappings
SQLite type mapping is extremely simple so this is mostly enabling the provider-agnostic tests for SQLite--which all work, even the ones that don't work on SQL Server!--and then adding SQLite specific tests that make sure that nothing barfs.

Note that size, precision, and scale are all ignored by SQLite, so tests for these just show that things still round-trip without issue.