Fix to #2069 - Guid.NewGuid() returns same order every time
Basically the issue is that we used to always attempt to pre-evaluate every method call and property that was not bound to the query (i.e. taking a column as an argument, or where column in an instance of the method/property).
This was generally desirable because users could inline some helper methods into their queries if they needed, rather than have to evaluate it manually outside the query and pass it as closure.
However, this is incorrect for functions that are not deterministic (e.g. Guid.NewGuid(), Datetime.Now). Those functions should always be evaluated on the server.
Users should be able to extend this list of non-deterministic functions (or rather functions that should never be pre-evaluated on the client) to include their own methods (TODO)

Fix is to provide this functionality. On top of that translation for many other BCL functions have been added. Before function that was bound to the query would be evaluated on the client after the information from database was obtained.
This could lead to severe perf issues, if such function was part of the predicate. With this change, function will be translated to corresponding SQL construct, so that entire predicate can be evaluated on the server, resulting less data returned to the client.

Function evaluation is performed by Relinq, and there is no good way to hook into the process, so what we do is temporarily wrap appropriate MethodCallExpressions and MemberExpresions with another expression that prevents evaluation of it's contents.
This way wrapped functions will be preserved and can be passed further into the query pipeline. During sql translation we recognize the methods that we know translations for and translate to appropriate SqlFunctionExpression. Functions that we don't know how to translate will behave as all functions used to before this change - i.e. will be evaluated on the client once all the data is received from the database.

CR: Andrew, Arthur