Fix to #2246 - Assigning boolean in Select statement throws invalid cast
Problem is that a row projection containing a constant bool gets translated into integer (0 or 1) since 1 and 0 are the literal bit values for true and false. When we try to materialize, we try to cast integer 0 or 1 into bool, which causes an error.
Fix is to not send standalone constants to the store, but instead deal with them on the client. This is consistent with our approach to functions (funcletize as much as we can)