Use extension method to clean up IAccessor use
This change uses a couple of extension methods to allow access to underlying services without having to cast to IAccessor<TService> every time. It means that this kind of low-level access is available and discoverable whenever the .Infrastructure namespace is imported, but is otherwise still hidden in normal apps.

Top level objects (DbContext, DbSet, Database) that expose IAccessor<IServiceProvider> allow use of the second extension method overload to automatically go through the service provider indirection and obtain any service registerd in DI.