Fix to #2274 - Chaining multiple Include with ThenInclude throws System.InvalidOperationException
Problem is that for complex queries with multiple includes we were attempting to materialize entities based on incorrect readers. Basically the way it works is that when we process include we build a factories that produce entities from actively opened readers.

Those were being build in isolation, starting fresh for each IncludeSpecification. Problem is that multiple include chains can leave readers open, that the other include chain was not aware of. This would produce errors like NRE, index outside bounds, invalid cast etc, when attempting to access incorrect reader while materializing entities.

Fix is to correctly compute indexes (which reader should be accessed for each entity) that each include chain needs to access. Computation is being done in the reverse order because this is the order that readers will eventually be executed in. Also we only compute indexes for Reference includes - Collections don't need them as they always open a new reader.

CR: Andrew