Fix to #2365 - ArgumentNullException when querying with nullable byte, short, or enum parameters
Problem was that in some cases of column == variable, compiler wraps the column expression into a Convert. As later stage in the query pipeline we try to recognize the pattern where column == null_valued_parameter and convert it to IsNull(column). However, we were not detecting it properly for columns wrapped in Convert. Fix is to unwrap converts in the pattern detection logic. We don't really need the information about the converted type, since the result expression is either the original (we didn't find the right pattern) or IsNull(column), that returns save boolean value regardless of the cast. CR: Andrew