Fix to #2554 - Relational: Incorrect column binding with the same column name
Problem was that we incorrectly assumed that join body clause will be placed directly after the main clause in the QueryModel. In fact, body clauses are added to the collection in the order in which the query was built. This caused us to be unable to find the previous query source (which we want to join with) and therefore we would not be able to produce a join query on the server.

Fix is to go through list of previously processed body clauses until we find a previous query source, rather than just checking one (previous) body clause.
CR: Andrew