Make optimistic concurrency usable
Introduces a new flag on property indicating that store generated values should always be used regardless of the current value in the property or whether or not it is marked as modified. By default any property that is a concurrency token and generated on add and update will also get this flag set, although it can be switched off. This makes rowversion optimistic concurrency just work out of the box.