:wrench: Write perf results to diagnostics
This allows us to switch on diagnostics and see the perf results in VS
output window or on the command line. This makes it easy to use perf
tests as a harness when working on perf improvements.
Also adding config files so that the diagnostic messages can be easily
switched on.