Fix to #2181 - Include tries to use PK rather than alternate key
Issue was that when building include expressions, we always assumed that joins should happen based on primary keys. However users can specify alternate key, and for those cases it should be used instead. Fix is to use correct navigation.ForeignKey.PrincipalKey to determine properties that should be used to create a join.
CR: Smit, Andrew