:bar_chart: Reporting for benchmark results
This is a pragmatic start at some reporting for benchmark results. There
is certainly lots of room for improvement (adopting a plugin to render
pivot tables, ajax loading of data for charts, etc.) but we can evolve
as we work out what information is useful to us.

The site will be hosted hooked up to our performance CI results, but you
can also just run it locally and it will look at the same local database
used to store local benchmark runs (which should be useful when working
on local perf improvements).