Fix to 2274 - Chaining multiple Include with ThenInclude throws in some cases when included collections are empty mid-chain
Problem is that when creating creating query plan for include we statically assign each index for each ReferenceEntity materializer that indicates from which reader the data should be taken, in order to materialize this entity.
However, in some cases, we don't open all the readers that we expect - this can happen if a reader doesn't return any results. We then short circuit the include chain, and don't open any readers that were suppose to follow.

Fix is to introduce a structure that holds active readers for include (on top of the old _activeQueries for all other readers), which we don't fill in like a list, but fill it based on the index that the reader is supposed to have (filling the blanks with nulls temporarily). This way we ensure that reader always land in their place, and are available at the appropriate index, that we introduced in the query plan. This happens whether all of them were opened, or some were short-circuited.

The information about which reader should be placed at which index is also added to the query plan. We do this only for the include-related readers.
CR: Andrew