Improve perf for snapshotting original values
In the common case where original values are tracked eagerly for all properties this change takes the query value buffer and uses it directly as the original values store. This essentially eliminates the cost of creating the original values snapshot, although snapshot of relationship state still takes time.

In doing this work it became apparent that sometimes values in the ValueBuffer are not of the correct type--notably for enum values. This caused problems for DetectChanges since the values were of different types and therefore were flagged as different even when no change had been made. To fix this the ValueBuffer obtained from GetValues is now post-processed to convert values as necessary. Currently this is only for enums, but in the future it could be used for any type conversion. Since values in the ValueBufferare now always of the correct type even when GetValues is used it means that the conversion code in the materializer can be removed.